[yixuan-refactor] output sinks
